Begynner å implementerer noen use-case

+++ application/src/main/kotlin/no/nilsen/cleana/salg/command/OpprettSalgDto.kt
@@ -1,5 +1,5 @@
 package no.nilsen.cleana.salg.command
 
-class OpprettSalgDto(val produktId: Int, val ansattId: Int, val antall: Int, val kundeId: Int) {
+class OpprettSalgDto(val produktId: Int, val ansattId: Int, val antall: Long, val kundeId: Int) {
 
 }
